Refactor update_dependencies
Move modules related to dependency updates into own folder.
