+++ arbeitszeit_development/update_dependencies/update_bulma.py
@@ -48,7 +48,12 @@ class BulmaUpdater:
 
     def get_target_path(self) -> Path:
         module_path = Path(__file__)
-        return module_path.parent.parent / "arbeitszeit_flask" / "static" / "bulma.css"
+        return (
+            module_path.parent.parent.parent
+            / "arbeitszeit_flask"
+            / "static"
+            / "bulma.css"
+        )
 
     def get_asset_name(self, release: GithubRelease) -> str:
         for asset in release.asset_names:
